+++ ikuzo/options.go
@@ -17,27 +17,14 @@ package ikuzo
 import (
 	"io/fs"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
-	"strings"
 
 	"github.com/delving/hub3/config"
+	"github.com/delving/hub3/ikuzo/domain"
 	"github.com/delving/hub3/ikuzo/logger"
 	"github.com/delving/hub3/ikuzo/service/organization"
-	"github.com/delving/hub3/ikuzo/service/x/bulk"
-	"github.com/delving/hub3/ikuzo/service/x/ead"
-	"github.com/delving/hub3/ikuzo/service/x/imageproxy"
-	"github.com/delving/hub3/ikuzo/service/x/oaipmh"
-	"github.com/delving/hub3/ikuzo/service/x/revision"
-	"github.com/delving/hub3/ikuzo/storage/x/elasticsearch"
 	"github.com/delving/hub3/ikuzo/webapp"
 	"github.com/go-chi/chi"
 	"github.com/pacedotdev/oto/otohttp"
-)
-
-const (
-	taskIDRoute    = "/api/ead/tasks/{id}"
-	datasetIDRoute = "/api/datasets/{spec}"
 )
 
 // RouterFunc is a callback that registers routes to the ikuzo.Server.
@@ -53,6 +40,14 @@ type Option func(*server) error
 func SetPort(port int) Option {
 	return func(s *server) error {
 		s.port = port
+		return nil
+	}
+}
+
+// RegisterService registers a Service with the ikuzo server
+func RegisterService(svc domain.Service) Option {
+	return func(s *server) error {
+		s.services = append(s.services, svc)
 		return nil
 	}
 }
@@ -122,53 +117,11 @@ func RegisterOtoServer(otoServer *otohttp.Server) Option {
 }
 
 // SetOrganisationService configures the organization service.
-// When no service is set a default transient memory-based service is used.
-func SetOrganisationService(service *organization.Service) Option {
+func SetOrganisationService(svc *organization.Service) Option {
 	return func(s *server) error {
-		s.organizations = service
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				r.Mount("/organizations", service.Routes())
-			},
-		)
-		s.middleware = append(s.middleware, service.ResolveOrgByDomain)
-
-		return nil
-	}
-}
-
-// SetRevisionService configures the organization service.
-// When no service is set a default transient memory-based service is used.
-func SetRevisionService(service *revision.Service) Option {
-	return func(s *server) error {
-		s.revision = service
-		s.routerFuncs = append(s.routerFuncs, func(r chi.Router) {
-			r.HandleFunc("/git/{user}/{collection}.git/*", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				p := strings.TrimPrefix(r.URL.Path, "/git")
-				if !service.BareRepo {
-					p = strings.ReplaceAll(p, ".git/", "/.git/")
-				}
-				r2 := new(http.Request)
-				*r2 = *r
-				r2.URL = new(url.URL)
-				*r2.URL = *r.URL
-				r2.URL.Path = p
-				service.ServeHTTP(w, r2)
-			}))
-		})
-
-		return nil
-	}
-}
-
-func SetElasticSearchProxy(proxy *elasticsearch.Proxy) Option {
-	return func(s *server) error {
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				r.Handle("/{index}/_search", proxy)
-				r.Handle("/{index}/{documentType}/_search", proxy)
-			},
-		)
+		s.organizations = svc
+		s.services = append(s.services, svc)
+		s.middleware = append(s.middleware, svc.ResolveOrgByDomain)
 
 		return nil
 	}
@@ -183,6 +136,35 @@ func SetBuildVersionInfo(info *BuildVersionInfo) Option {
 				})
 			},
 		)
+
+		return nil
+	}
+}
+
+// SetStaticFS registers an fs.FS as a static fileserver.
+//
+// It is mounts '/static/*' and '/favicon.ico'.
+// Note: it can only be set once. So you can register multiple fs.FS.
+func SetStaticFS(static fs.FS) Option {
+	return func(s *server) error {
+		s.routerFuncs = append(s.routerFuncs,
+			func(r chi.Router) {
+				r.Get("/static/*", webapp.NewStaticHandler(static))
+				r.Get("/favicon.ico", func(w http.ResponseWriter, r *http.Request) {
+					http.Redirect(w, r, "/static/favicon.ico", http.StatusMovedPermanently)
+				})
+			},
+		)
+
+		return nil
+	}
+}
+
+func SetShutdownHook(name string, hook Shutdown) Option {
+	return func(s *server) error {
+		if _, ok := s.shutdownHooks[name]; !ok {
+			s.shutdownHooks[name] = hook
+		}
 
 		return nil
 	}
@@ -207,140 +189,9 @@ func SetLegacyRouters(routers ...RouterFunc) Option {
 	}
 }
 
-func SetStaticFS(static fs.FS) Option {
+func SetEnableIntrospect(enabled bool) Option {
 	return func(s *server) error {
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				r.Get("/static/*", webapp.NewStaticHandler(static))
-				r.Get("/favicon.ico", func(w http.ResponseWriter, r *http.Request) {
-					http.Redirect(w, r, "/static/favicon.ico", http.StatusMovedPermanently)
-				})
-			},
-		)
-
-		return nil
-	}
-}
-
-func SetEADService(svc *ead.Service) Option {
-	return func(s *server) error {
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				r.Post("/api/ead", svc.Upload)
-				r.Get("/api/ead/tasks", svc.Tasks)
-				r.Get(taskIDRoute, svc.GetTask)
-				r.Delete(taskIDRoute, svc.CancelTask)
-			},
-		)
-
-		s.addShutdown("EAD service", svc)
-
-		return nil
-	}
-}
-
-func SetOAIPMHService(svc *oaipmh.Service) Option {
-	return func(s *server) error {
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				r.Get("/oai/!open_oai.OAIHandler", svc.ServeHTTP)
-				r.Post("/oai/harvest-now", svc.HarvestNow)
-			},
-		)
-
-		s.addShutdown("oai-pmh service", svc)
-
-		return nil
-	}
-}
-
-func SetBulkService(svc *bulk.Service) Option {
-	return func(s *server) error {
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				r.Post("/api/index/bulk", svc.Handle)
-				r.Post("/api/index/rdf", svc.HandleRDF)
-			},
-		)
-
-		return nil
-	}
-}
-
-func SetShutdownHook(name string, hook Shutdown) Option {
-	return func(s *server) error {
-		if _, ok := s.shutdownHooks[name]; !ok {
-			s.shutdownHooks[name] = hook
-		}
-
-		return nil
-	}
-}
-
-func SetImageProxyService(service *imageproxy.Service) Option {
-	return func(s *server) error {
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				service.Routes("", r)
-			},
-		)
-
-		return nil
-	}
-}
-
-type ProxyRoute struct {
-	Method  string
-	Pattern string
-}
-
-// SetDataNodeProxy creates a reverse proxy to the dataNode and set override routes.
-//
-// The 'proxyRoutes' argument can be used to add additional override routes.
-func SetDataNodeProxy(dataNode string, proxyRoutes ...ProxyRoute) Option {
-	return func(s *server) error {
-		nodeURL, _ := url.Parse(dataNode)
-		s.dataNodeProxy = httputil.NewSingleHostReverseProxy(nodeURL)
-		s.routerFuncs = append(s.routerFuncs,
-			func(r chi.Router) {
-				// ead
-				r.Post("/api/ead", s.proxyDataNode)
-				r.Get("/api/ead/tasks", s.proxyDataNode)
-				r.Get(taskIDRoute, s.proxyDataNode)
-				r.Delete(taskIDRoute, s.proxyDataNode)
-				r.Post("/api/index/bulk", s.proxyDataNode)
-				r.Get("/api/ead/{spec}/download", s.proxyDataNode)
-				r.Get("/api/ead/{spec}/mets/{inventoryID}", s.proxyDataNode)
-				r.Get("/api/ead/{spec}/desc", s.proxyDataNode)
-				r.Get("/api/ead/{spec}/desc/index", s.proxyDataNode)
-				r.Get("/api/ead/{spec}/meta", s.proxyDataNode)
-
-				// datasets
-				r.Get("/api/datasets/", s.proxyDataNode)
-				r.Get("/api/datasets/histogram", s.proxyDataNode)
-				r.Post("/api/datasets/", s.proxyDataNode)
-				r.Get(datasetIDRoute, s.proxyDataNode)
-				r.Get("/api/datasets/{spec}/stats", s.proxyDataNode)
-				// later change to update dataset
-				r.Post(datasetIDRoute, s.proxyDataNode)
-				r.Delete(datasetIDRoute, s.proxyDataNode)
-
-				// custom routes
-				for _, route := range proxyRoutes {
-					switch {
-					case strings.EqualFold("get", route.Method):
-						r.Get(route.Pattern, s.proxyDataNode)
-					case strings.EqualFold("post", route.Method):
-						r.Post(route.Pattern, s.proxyDataNode)
-					case strings.EqualFold("put", route.Method):
-						r.Put(route.Pattern, s.proxyDataNode)
-					case strings.EqualFold("delete", route.Method):
-						r.Delete(route.Pattern, s.proxyDataNode)
-					}
-				}
-			},
-		)
-
+		s.introspect = enabled
 		return nil
 	}
 }