+++ ikuzo/service/x/ead/service.go
@@ -35,15 +35,14 @@ import (
 	"github.com/olivere/elastic/v7"
 	"github.com/rs/zerolog"
 
-	c "github.com/delving/hub3/config"
 	eadHub3 "github.com/delving/hub3/hub3/ead"
 	"github.com/delving/hub3/hub3/fragments"
 	indexHub3 "github.com/delving/hub3/hub3/index"
 	"github.com/delving/hub3/hub3/models"
 	"github.com/delving/hub3/ikuzo/domain"
+	"github.com/delving/hub3/ikuzo/driver/elasticsearch"
 	"github.com/delving/hub3/ikuzo/service/x/index"
 	"github.com/delving/hub3/ikuzo/service/x/oaipmh"
-	"github.com/delving/hub3/ikuzo/service/x/revision"
 	"github.com/go-chi/chi"
 	"github.com/go-chi/render"
 	"github.com/rs/zerolog/log"
@@ -54,48 +53,16 @@ const (
 	PaccessKey = "processAccessTime"
 )
 
-type Metrics struct {
-	Submitted     uint64
-	Started       uint64
-	Failed        uint64
-	Finished      uint64
-	Canceled      uint64
-	AlreadyQueued uint64
-}
-
-func (m *Metrics) IncSubmitted() {
-	atomic.AddUint64(&m.Submitted, 1)
-}
-
-func (m *Metrics) IncStarted() {
-	atomic.AddUint64(&m.Started, 1)
-}
-
-func (m *Metrics) IncFailed() {
-	atomic.AddUint64(&m.Failed, 1)
-}
-
-func (m *Metrics) IncFinished() {
-	atomic.AddUint64(&m.Finished, 1)
-}
-
-func (m *Metrics) IncCancelled() {
-	atomic.AddUint64(&m.Canceled, 1)
-}
-
-func (m *Metrics) IncAlreadyQueued() {
-	atomic.AddUint64(&m.AlreadyQueued, 1)
-}
-
 type CreateTreeFn func(cfg *eadHub3.NodeConfig, n *eadHub3.Node, hubID string, id string) *fragments.Tree
 
 type PreStoreFn func(b []byte) []byte
 
 type DaoFn func(cfg *eadHub3.DaoConfig) error
 
+var _ domain.Service = (*Service)(nil)
+
 type Service struct {
 	index                   *index.Service
-	revision                *revision.Service
 	dataDir                 string
 	M                       Metrics
 	CreateTreeFn            CreateTreeFn
@@ -110,6 +77,8 @@ type Service struct {
 	cancel                  context.CancelFunc
 	group                   *errgroup.Group
 	postHooks               map[string][]domain.PostHookService
+	log                     zerolog.Logger
+	orgs                    domain.OrgConfigRetriever
 }
 
 func NewService(options ...Option) (*Service, error) {
@@ -154,26 +123,15 @@ func NewService(options ...Option) (*Service, error) {
 	return s, nil
 }
 
-func (s *Service) findAvailableTask() *Task {
-	tasks := []*Task{}
+func (s *Service) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	router := chi.NewRouter()
+	s.Routes("", router)
+	router.ServeHTTP(w, r)
+}
 
-	for _, task := range s.tasks {
-		if task.InState == StatePending || task.Interrupted {
-			tasks = append(tasks, task)
-		}
-	}
-
-	if len(tasks) == 0 {
-		return nil
-	}
-
-	sort.Slice(tasks, func(i, j int) bool {
-		return tasks[i].currentTransition().Started.After(tasks[j].currentTransition().Started)
-	})
-
-	log.Info().Str("svc", "eadProcessor").Int("availableTasks", len(tasks)).Msg("returning first available task for processing")
-
-	return tasks[0]
+func (s *Service) SetServiceBuilder(b *domain.ServiceBuilder) {
+	s.log = b.Logger.With().Str("svc", "sitemap").Logger()
+	s.orgs = b.Orgs
 }
 
 func (s *Service) StartWorkers() error {
@@ -231,74 +189,6 @@ func (s *Service) Upload(w http.ResponseWriter, r *http.Request) {
 // Call this function each night at 00:01 in a cron job to check and clear tree node restrictions.
 func (s *Service) ClearRestrictions(w http.ResponseWriter, r *http.Request) {
 	s.clearRestrictions(w, r)
-}
-
-func (s *Service) Tasks(w http.ResponseWriter, r *http.Request) {
-	s.rw.RLock()
-	defer s.rw.RUnlock()
-
-	// TODO(kiivihal): add option to filter by datasetID
-
-	render.JSON(w, r, s.tasks)
-}
-
-func (s *Service) findTask(orgID, datasetID string, filterActive bool) (*Task, error) {
-	s.rw.RLock()
-	defer s.rw.RUnlock()
-
-	for _, t := range s.tasks {
-		// TODO(kiivihal): add filter for orgID later
-		_ = orgID
-
-		if t.Meta.DatasetID == datasetID {
-			if filterActive && !t.isActive() {
-				continue
-			}
-
-			return t, nil
-		}
-	}
-
-	return nil, ErrTaskNotFound
-}
-
-func (s *Service) GetTask(w http.ResponseWriter, r *http.Request) {
-	s.rw.RLock()
-	defer s.rw.RUnlock()
-
-	id := chi.URLParam(r, "id")
-
-	task, ok := s.tasks[id]
-	if !ok {
-		http.Error(w, "unknown task", http.StatusNotFound)
-		return
-	}
-
-	render.JSON(w, r, task)
-}
-
-func (s *Service) CancelTask(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-
-	s.rw.Lock()
-	defer s.rw.Unlock()
-
-	task, ok := s.tasks[id]
-	if !ok {
-		http.Error(w, "unknown task", http.StatusNotFound)
-		return
-	}
-
-	task.moveState(StateCanceled)
-
-	task.log().Info().Msg("canceling running ead task")
-	task.cancel()
-
-	task.Next()
-	// TODO(kiivihal): do we delete or keep it
-	// delete(s.tasks, id)
-
-	w.WriteHeader(http.StatusNoContent)
 }
 
 func (s *Service) Shutdown(ctx context.Context) error {
@@ -759,9 +649,10 @@ func (s *Service) CreateTask(r *http.Request, meta Meta) (*taskResponse, error) 
 
 // TODO: move this to elasticsearch package later
 func (s *Service) clearRestrictions(w http.ResponseWriter, r *http.Request) {
-	orgID := domain.GetOrganizationID(r)
+	org, _ := domain.GetOrganization(r)
 
-	search := indexHub3.ESClient().Search(c.Config.ElasticSearch.GetIndexName())
+	search := indexHub3.ESClient().Search(org.Config.GetIndexName())
+
 	format := "02-01-2006"
 	today := time.Now()
 	if r.URL.Query().Get(PaccessKey) != "" {
@@ -774,7 +665,7 @@ func (s *Service) clearRestrictions(w http.ResponseWriter, r *http.Request) {
 	}
 	query := elastic.NewBoolQuery().
 		Filter(
-			elastic.NewMatchPhraseQuery("meta.orgID", orgID.String()),
+			elastic.NewMatchPhraseQuery(elasticsearch.PathOrgID, org.RawID()),
 			elastic.NewMatchPhraseQuery("tree.hasRestriction", true),
 			elastic.NewMatchPhraseQuery("tree.type", "file"),
 			elastic.NewMatchPhraseQuery("tree.access", today.Format(format)),
@@ -808,7 +699,7 @@ func (s *Service) clearRestrictions(w http.ResponseWriter, r *http.Request) {
 	trSlice := make([]*taskResponse, 0)
 	for _, el := range terms.Buckets {
 		spec := el.Key.(string)
-		meta, err := s.LoadEAD(orgID.String(), spec)
+		meta, err := s.LoadEAD(org.RawID(), spec)
 		if err != nil {
 			clearLogger.Err(err).Msgf("could not load spec %s from bucket: %v", spec, err)
 			continue
@@ -1026,13 +917,6 @@ func (s *Service) AddPostHook(hook domain.PostHookService) error {
 	}
 
 	return nil
-}
-
-func (s *Service) Routes(router chi.Router) {
-	router.Get("/api/ead/{spec}/mets/{UUID}.json", s.DaoClient.DownloadConfig)
-	router.Get("/api/ead/{spec}/mets/{UUID}", s.DaoClient.DownloadXML)
-	router.Delete("/api/ead/{spec}/mets/{UUID}", s.DaoClient.HandleDelete)
-	router.Post("/api/ead/{spec}/mets/{UUID}", s.DaoClient.Index)
 }
 
 // getModifiedDatePath returns the filepath of the EAD modified date.